+++ templates/cra-template-lattice-apollo-enterprise/template/src/containers/PrivateRoute.js
@@ -1,0 +1,30 @@
+import React from 'react';
+
+import { Redirect, Route } from 'react-router-dom';
+
+import { withCurrentUser } from './Auth';
+
+import { SIGN_IN } from './routes';
+
+const PrivateRoute = (props) => {
+  const { currentUser, component: RouteComponent, render, ...rest } = props;
+  return (
+    <Route
+      {...rest}
+      render={(props) => {
+        const { location } = props;
+        return currentUser ? (
+          RouteComponent ? (
+            <RouteComponent {...props} currentUser={currentUser} />
+          ) : (
+            render(props)
+          )
+        ) : (
+          <Redirect to={{ pathname: SIGN_IN, state: { from: location } }} />
+        );
+      }}
+    />
+  );
+};
+
+export default withCurrentUser(PrivateRoute);