+++ templates/cra-template-lattice-apollo-enterprise/template/src/stores/task.js
@@ -1,0 +1,19 @@
+import gql from 'graphql-tag';
+
+export const tasksConnection = gql`
+  query TasksConnection($filterBy: [Filter], $orderBy: [Order], $first: Int = 10, $after: String) {
+    tasksConnection(filterBy: $filterBy, orderBy: $orderBy, first: $first, after: $after) {
+      totalCount
+      edges {
+        node {
+          title
+        }
+        cursor
+      }
+      pageInfo {
+        endCursor
+        hasNextPage
+      }
+    }
+  }
+`;